refactor(Topbar): Extract EditableAlbumName into its own file
